Move `isatty` into the termios API module.
